changed mappers organization and add request dto and mapping

+++ house-seeker-domain/src/main/java/pl/kala/houseseekerdomain/database/model/document/house/House.java
@@ -46,7 +46,7 @@ public class House {
     private LocalDate buildDate;
 
     @NonNull
-    private float pricePerSqMeter;
+    private double pricePerSqMeter;
 
     private boolean hasBalcony;
